refactor(architecture): organize contacts introduction flow
