+++ features/flow/flow-contacts-introduction/README.md
@@ -11,6 +11,15 @@ Contacts Feature Introduction and Ledger Sync Introduction leaf flow for Desktop
 - Ledger Sync Introduction status and resolver.
 - Web dialogs and React Native bottom-sheet content for both introductions.
 
+## Structure
+
+- `screens/FeatureIntroduction/` owns the Feature Introduction rendering, assets, and actions.
+- `screens/LedgerSyncIntroduction/` owns the Ledger Sync Introduction rendering.
+- `state/` owns the resolver, preference port, state hook, constants, and shared introduction types.
+
+Components stay colocated with the screen that owns them. A root `components/` directory is
+reserved for components genuinely shared by both introduction screens.
+
 ## Public API
 
 Apps that mount an introduction directly import this package for its hooks, types, helpers, and
