+++ features/flow/flow-contacts-introduction/src/screens/FeatureIntroduction/ContactsFeatureIntroductionDialog.web.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { Button, Dialog, DialogBody, DialogContent, DialogHeader } from "@ledgerhq/lumen-ui-react";
-import type { ContactsFeatureIntroduction } from "../types";
-import { useContactsFeatureIntroductionActions } from "../useContactsFeatureIntroductionActions";
+import type { ContactsFeatureIntroduction } from "../../state/types";
+import { useContactsFeatureIntroductionActions } from "./useContactsFeatureIntroductionActions";
 import { CONTACTS_FEATURE_INTRODUCTION_HERO_IMAGE } from "./assets";
 import { ContactsFeatureIntroductionDialogContent } from "./ContactsFeatureIntroductionDialogContent.web";
